14-Clover update to Zig 0.14 and contemporaneous zigwin32

+++ Chapter 05 Basic Drawing/14-Clover/Clover.zig
@@ -16,12 +16,11 @@ const std = @import("std");
 const WINAPI = std.os.windows.WINAPI;
 
 const win32 = struct {
-    usingnamespace @import("win32").zig;
-    usingnamespace @import("win32").system.library_loader;
-    usingnamespace @import("win32").foundation;
-    usingnamespace @import("win32").system.system_services;
-    usingnamespace @import("win32").ui.windows_and_messaging;
-    usingnamespace @import("win32").graphics.gdi;
+    usingnamespace @import("zigwin32").zig;
+    usingnamespace @import("zigwin32").system.library_loader;
+    usingnamespace @import("zigwin32").foundation;
+    usingnamespace @import("zigwin32").ui.windows_and_messaging;
+    usingnamespace @import("zigwin32").graphics.gdi;
 };
 const BOOL = win32.BOOL;
 const TRUE = win32.TRUE;
@@ -30,23 +29,13 @@ const L = win32.L;
 const HINSTANCE = win32.HINSTANCE;
 const MSG = win32.MSG;
 const HWND = win32.HWND;
-const HDC = win32.HDC;
 const LPARAM = win32.LPARAM;
 const WPARAM = win32.WPARAM;
 const LRESULT = win32.LRESULT;
 const HRGN = win32.HRGN;
-const CW_USEDEFAULT = win32.CW_USEDEFAULT;
-const WS_OVERLAPPEDWINDOW = win32.WS_OVERLAPPEDWINDOW;
-const WM_SIZE = win32.WM_SIZE;
-const WM_PAINT = win32.WM_PAINT;
-const WM_DESTROY = win32.WM_DESTROY;
 
-const windowsx = @import("windowsx").windowsx;
+const windowsx = @import("windowsx");
 const GetStockBrush = windowsx.GetStockBrush;
-const SelectBrush = windowsx.SelectBrush;
-const HANDLE_WM_SIZE = windowsx.HANDLE_WM_SIZE;
-const HANDLE_WM_PAINT = windowsx.HANDLE_WM_PAINT;
-const HANDLE_WM_DESTROY = windowsx.HANDLE_WM_DESTROY;
 
 pub export fn wWinMain(
     hInstance: HINSTANCE,
@@ -57,13 +46,13 @@ pub export fn wWinMain(
     const app_name = L("Clover");
     const wndclassex = win32.WNDCLASSEX{
         .cbSize = @sizeOf(win32.WNDCLASSEX),
-        .style = win32.WNDCLASS_STYLES.initFlags(.{ .HREDRAW = 1, .VREDRAW = 1 }),
+        .style = win32.WNDCLASS_STYLES{ .HREDRAW = 1, .VREDRAW = 1 },
         .lpfnWndProc = WndProc,
         .cbClsExtra = 0,
         .cbWndExtra = 0,
         .hInstance = hInstance,
-        .hIcon = @ptrCast(win32.HICON, win32.LoadImage(null, win32.IDI_APPLICATION, win32.IMAGE_ICON, 0, 0, win32.IMAGE_FLAGS.initFlags(.{ .SHARED = 1, .DEFAULTSIZE = 1 }))),
-        .hCursor = @ptrCast(win32.HCURSOR, win32.LoadImage(null, win32.IDC_ARROW, win32.IMAGE_CURSOR, 0, 0, win32.IMAGE_FLAGS.initFlags(.{ .SHARED = 1, .DEFAULTSIZE = 1 }))),
+        .hIcon = @ptrCast(win32.LoadImage(null, win32.IDI_APPLICATION, win32.IMAGE_ICON, 0, 0, win32.IMAGE_FLAGS{ .SHARED = 1, .DEFAULTSIZE = 1 })),
+        .hCursor = @ptrCast(win32.LoadImage(null, win32.IDC_ARROW, win32.IMAGE_CURSOR, 0, 0, win32.IMAGE_FLAGS{ .SHARED = 1, .DEFAULTSIZE = 1 })),
         .hbrBackground = GetStockBrush(win32.WHITE_BRUSH),
         .lpszMenuName = null,
         .lpszClassName = app_name,
@@ -72,27 +61,23 @@ pub export fn wWinMain(
 
     const atom: u16 = win32.RegisterClassEx(&wndclassex);
     if (0 == atom) {
-        std.debug.print("failed RegisterClassEx()", .{});
+        std.log.err("failed RegisterClassEx()", .{});
         return 0; // premature exit
     }
 
-    // If a memory align panic occurs then the CreateWindowExW() Zig declaration
-    // needs to have align(1) added to the lpClassName parameter.
-    //   lpClassName: ?[*:0]align(1) const u16,
-    //                      ^^^^^^^^
+    // If a memory align panic occurs with CreateWindowExW() lpClassName then look at:
     // https://github.com/marlersoft/zigwin32gen/issues/9
-    const lpClassName = @intToPtr([*:0]align(1) const u16, atom);
 
     const hwnd = win32.CreateWindowEx(
         // https://docs.microsoft.com/en-us/windows/win32/winmsg/extended-window-styles
-        win32.WINDOW_EX_STYLE.initFlags(.{}),
-        lpClassName,
+        win32.WINDOW_EX_STYLE{},
+        @ptrFromInt(atom),
         L("Draw a Clover"),
-        WS_OVERLAPPEDWINDOW,
-        CW_USEDEFAULT, // initial x position
-        CW_USEDEFAULT, // initial y position
-        CW_USEDEFAULT, // initial x size
-        CW_USEDEFAULT, // initial y size
+        win32.WS_OVERLAPPEDWINDOW,
+        win32.CW_USEDEFAULT, // initial x position
+        win32.CW_USEDEFAULT, // initial y position
+        win32.CW_USEDEFAULT, // initial x size
+        win32.CW_USEDEFAULT, // initial y size
         null, // parent window handle
         null, // window menu handle
         win32.GetModuleHandle(null),
@@ -100,13 +85,13 @@ pub export fn wWinMain(
     );
 
     if (null == hwnd) {
-        std.debug.print("failed CreateWindowEx(), error {}", .{win32.GetLastError()});
+        std.log.err("failed CreateWindowEx(), error {}", .{win32.GetLastError()});
         return 0; // premature exit
     }
 
-    _ = win32.ShowWindow(hwnd, @intToEnum(win32.SHOW_WINDOW_CMD, nCmdShow));
+    _ = win32.ShowWindow(hwnd, @bitCast(nCmdShow));
     if (0 == win32.UpdateWindow(hwnd)) {
-        std.debug.print("failed UpdateWindow()", .{});
+        std.log.err("failed UpdateWindow()", .{});
         return 0; // premature exit
     }
 
@@ -117,7 +102,7 @@ pub export fn wWinMain(
         if (-1 == ret) {
             // handle the error and/or exit
             // for error call GetLastError();
-            std.debug.print("failed message loop, error {}", .{win32.GetLastError()});
+            std.log.err("failed message loop, error {}", .{win32.GetLastError()});
             return 0;
         } else {
             _ = win32.TranslateMessage(&msg);
@@ -127,7 +112,7 @@ pub export fn wWinMain(
     }
 
     // Normal exit
-    return @bitCast(c_int, @truncate(c_uint, msg.wParam)); // WM_QUIT
+    return @bitCast(@as(c_uint, @truncate(msg.wParam))); // WM_QUIT
 }
 
 const Handler = struct {
@@ -139,12 +124,20 @@ const Handler = struct {
         self.cxClient = cx;
         self.cyClient = cy;
 
-        const waitCursor = @ptrCast(
-            win32.HCURSOR,
-            win32.LoadImage(null, win32.IDC_WAIT, win32.IMAGE_CURSOR, 0, 0, win32.IMAGE_FLAGS.initFlags(.{ .SHARED = 1, .DEFAULTSIZE = 1 })),
-        );
+        const waitCursor: win32.HCURSOR = @ptrCast(win32.LoadImage(
+            null,
+            win32.IDC_WAIT,
+            win32.IMAGE_CURSOR,
+            0,
+            0,
+            win32.IMAGE_FLAGS{ .SHARED = 1, .DEFAULTSIZE = 1 },
+        ));
+
         const previousCursor = win32.SetCursor(waitCursor);
+        defer _ = win32.SetCursor(previousCursor);
+
         _ = win32.ShowCursor(TRUE);
+        defer _ = win32.ShowCursor(FALSE);
 
         if (self.hRgnClip) |hRgn|
             _ = win32.DeleteObject(hRgn);
@@ -183,11 +176,11 @@ const Handler = struct {
         _ = win32.CombineRgn(hRgnTemp[5], hRgnTemp[2], hRgnTemp[3], win32.RGN_OR);
         _ = win32.CombineRgn(self.hRgnClip, hRgnTemp[4], hRgnTemp[5], win32.RGN_XOR);
 
-        for (hRgnTemp) |hRgn|
-            _ = win32.DeleteObject(hRgn);
-
-        _ = win32.SetCursor(previousCursor);
-        _ = win32.ShowCursor(FALSE);
+        for (hRgnTemp) |optional| {
+            if (optional) |hRgn| {
+                _ = win32.DeleteObject(hRgn);
+            }
+        }
     }
 
     pub fn OnPaint(self: *Handler, hwnd: HWND) void {
@@ -198,15 +191,15 @@ const Handler = struct {
         _ = win32.SetViewportOrgEx(hdc, @divTrunc(self.cxClient, 2), @divTrunc(self.cyClient, 2), null);
         _ = win32.SelectClipRgn(hdc, self.hRgnClip);
 
-        const fRadius = std.math.hypot(f32, @intToFloat(f32, self.cxClient) / 2.0, @intToFloat(f32, self.cyClient) / 2.0);
+        const fRadius = std.math.hypot(@as(f32, @floatFromInt(self.cxClient)) / 2.0, @as(f32, @floatFromInt(self.cyClient)) / 2.0);
         var fAngle: f32 = 0.0;
         const inc = 2.0 * std.math.pi / 360.0;
         while (fAngle < 2 * std.math.pi) : (fAngle += inc) {
             _ = win32.MoveToEx(hdc, 0, 0, null);
             _ = win32.LineTo(
                 hdc,
-                @floatToInt(i32, fRadius * std.math.cos(fAngle) + 0.5),
-                @floatToInt(i32, -fRadius * std.math.sin(fAngle) + 0.5),
+                @intFromFloat(fRadius * @cos(fAngle) + 0.5),
+                @intFromFloat(-fRadius * @sin(fAngle) + 0.5),
             );
         }
     }
@@ -216,13 +209,22 @@ const Handler = struct {
     }
 };
 
-var handler = Handler{};
+const WM_SIZE = win32.WM_SIZE;
+const WM_PAINT = win32.WM_PAINT;
+const WM_DESTROY = win32.WM_DESTROY;
+const HANDLE_WM_SIZE = windowsx.HANDLE_WM_SIZE;
+const HANDLE_WM_PAINT = windowsx.HANDLE_WM_PAINT;
+const HANDLE_WM_DESTROY = windowsx.HANDLE_WM_DESTROY;
 
 fn WndProc(hwnd: HWND, message: u32, wParam: WPARAM, lParam: LPARAM) callconv(WINAPI) LRESULT {
+    const state = struct {
+        var handler = Handler{};
+    };
+
     return switch (message) {
-        WM_SIZE => HANDLE_WM_SIZE(hwnd, wParam, lParam, Handler, &handler),
-        WM_PAINT => HANDLE_WM_PAINT(hwnd, wParam, lParam, Handler, &handler),
-        WM_DESTROY => HANDLE_WM_DESTROY(hwnd, wParam, lParam, Handler, &handler),
+        WM_SIZE => HANDLE_WM_SIZE(hwnd, wParam, lParam, Handler, &state.handler),
+        WM_PAINT => HANDLE_WM_PAINT(hwnd, wParam, lParam, Handler, &state.handler),
+        WM_DESTROY => HANDLE_WM_DESTROY(hwnd, wParam, lParam, Handler, &state.handler),
         else => win32.DefWindowProc(hwnd, message, wParam, lParam),
     };
 }
